Add some model,service and view changes

+++ Data/HomeBook.Data.Models/Apartment.cs
@@ -24,7 +24,7 @@
         [MaxLength(GlobalConstants.DataValidations.ApartmentResidentsMaxNumber)]
         public int NumberOfResidents { get; set; }
 
-        public decimal Area { get; set; }
+        public double Area { get; set; }
 
         public int EntranceId { get; set; }
